requirements.txt
streamlit>=1.33
sqlalchemy>=2.0
pandas>=2.2
numpy>=1.26
opencv-python>=4.9
scikit-learn>=1.4
scikit-image>=0.22
matplotlib>=3.8
seaborn>=0.13
Pillow>=10.3
tensorflow>=2.16
